fix: deep audit follow-up — IPC orphan rejection, Rust traversal guards, misc hardening
Solde de l'audit code 2026-06-09 (suites des passes 1-9, commit 41ecce9) :
- ipc.js : la branche invoke orpheline apres un timeout gagne n'emet
  plus d'unhandledrejection (catch noop)
- commands.rs : check_paths rejette les composants .. non canonicalises
  (traversal), clippy redundant_closure sur is_safe_dir
- cfg.js : constantes VIRT_ROW_H_COMPACT/COMFORTABLE pour tlistZoom
  (plus de hauteurs en dur)
- backup/cdaudio/dupes/playlog/watchfolder/sleep/modal/selection/
  shortcuts/playlist-*/renderer/cinema-viz/lf-toast-stack : gardes
  d'erreur et corrections diverses de l'audit

npm test 398/398, cargo test 9/9 (target-test, binaire app verrouille).

+++ src-tauri/src/taskbar.rs
@@ -178,8 +178,8 @@ fn rebuild_image_lists(old_play: isize, old_pause: isize) -> (isize, isize) {
             let _ = ImageList_Destroy(HIMAGELIST(old_pause));
         }
     }
-    let il_play_raw = build_image_list(false).0 as isize;
-    let il_pause_raw = build_image_list(true).0 as isize;
+    let il_play_raw = build_image_list(false).0;
+    let il_pause_raw = build_image_list(true).0;
     (il_play_raw, il_pause_raw)
 }
